refactor: move AddDebitController to a AddDebit folder
