--- src/presentation/controllers/debits/add-debit-protocols.ts
@@ -1,3 +1,0 @@
-export * from '../../protocols'
-export * from '../../../domain/models/debit'
-export * from '../../../domain/usecases/add-debit'
