+++ src/presentation/controllers/debits/add-debit/add-debit-controller.ts
@@ -1,6 +1,6 @@
 import { AddDebit, Controller, HttpResponse, HttpRequest } from './add-debit-protocols'
-import { serverError,ok } from '../../helpers/http/http-helpers'
-import { ServerError } from '../../errors'
+import { serverError,ok } from '../../../helpers/http/http-helpers'
+import { ServerError } from '../../../errors'
 export class AddDebitController implements Controller {
   constructor (private readonly addDebit: AddDebit) {}
   async handle (httpRequest: HttpRequest): Promise<HttpResponse> {
