+++ src/presentation/controllers/debits/add-debit/add-debit-protocols.ts
@@ -1,0 +1,3 @@
+export * from '../../../protocols'
+export * from '../../../../domain/models/debit'
+export * from '../../../../domain/usecases/add-debit'
